+++ cardigan/stories/components/Cards/FeaturedCard/FeaturedCard.stories.tsx
@@ -1,10 +1,8 @@
 import { Meta, StoryObj } from '@storybook/react';
 
 import { ReadmeDecorator } from '@weco/cardigan/config/decorators';
-import {
-  contentAPIArticle,
-  exhibitionBasic,
-} from '@weco/cardigan/stories/data/content';
+import { exhibitionBasic } from '@weco/cardigan/stories/data/content';
+import { contentAPIArticle } from '@weco/cardigan/stories/data/contentAPI.content';
 import { image } from '@weco/cardigan/stories/data/images';
 import { font } from '@weco/common/utils/classnames';
 import FeaturedCard from '@weco/content/views/components/FeaturedCard';
